requirements.txt
alembic==1.10.4
asyncpg==0.27.0
httpx==0.24.0
factory-boy==3.2.1
orjson==3.8.12
sqlmodel==0.0.8
fastapi==0.95.1
pytest==7.3.1
pytest-asyncio==0.21.0
pre-commit==3.3.1
gunicorn==20.1.0
uvicorn==0.22.0
